# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HRUO Enterprise HCM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
...
  /leave/types:
    get:
      operationId: listLeaveTypes
      tags: [leave]
      summary: List leave types
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveType"
    post:
      operationId: createLeaveType
      tags: [leave]
      summary: Create leave type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveTypeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveType"
  /leave/requests:
    get:
      operationId: listLeaveRequests
      tags: [leave]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leave]
      summary: Submit leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/requests/{id}/approve:
    post:
      operationId: approveLeaveRequest
      tags: [leave]
      summary: Approve leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/balances/{employeeId}:
    get:
      operationId: getLeaveBalances
      tags: [leave]
      summary: Get leave balances for employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveBalance"

  # ── ATTENDANCE ────────────────────────────────────────────────────────────
...
  /expenses/claims:
    get:
      operationId: listExpenseClaims
      tags: [expenses]
      summary: List expense claims
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseClaim"
    post:
      operationId: createExpenseClaim
      tags: [expenses]
      summary: Submit expense claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseClaimInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseClaim"
  /expenses/claims/{id}/approve:
    post:
      operationId: approveExpenseClaim
      tags: [expenses]
      summary: Approve or reject expense claim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalAction"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseClaim"
  /expenses/categories:
    get:
      operationId: listExpenseCategories
      tags: [expenses]
      summary: List expense categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseCategory"
    post:
      operationId: createExpenseCategory
      tags: [expenses]
      summary: Create expense category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseCategory"

  # ── PERFORMANCE ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, employeeNumber, firstName, lastName, email, status]
      properties:
        id:
          type: integer
        employeeNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        photo:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        nationalId:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        branchId:
          type: ["integer", "null"]
        branchName:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        dateJoined:
          type: ["string", "null"]
        probationEnd:
          type: ["string", "null"]
        employmentType:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, terminated, on_leave]
        basicSalary:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
    EmployeeInput:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gender:
          type: string
        birthDate:
          type: string
        nationality:
          type: string
        nationalId:
          type: string
        jobTitle:
          type: string
        departmentId:
          type: ["integer", "null"]
        branchId:
          type: ["integer", "null"]
        managerId:
          type: ["integer", "null"]
        dateJoined:
          type: string
        employmentType:
          type: string
        basicSalary:
          type: number
        currency:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        jobTitle:
          type: string
        departmentId:
          type: ["integer", "null"]
        branchId:
          type: ["integer", "null"]
        managerId:
          type: ["integer", "null"]
        status:
          type: string
        basicSalary:
          type: number
        bankName:
          type: string
        iban:
          type: string
...
    LeaveType:
      type: object
      required: [id, name, code, paid, active]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        color:
          type: ["string", "null"]
        paid:
          type: boolean
        active:
          type: boolean
        daysPerYear:
          type: ["number", "null"]
    LeaveTypeInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        color:
          type: string
        paid:
          type: boolean
        daysPerYear:
          type: number
    LeaveRequest:
      type: object
      required: [id, employeeId, leaveTypeId, startDate, endDate, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        leaveTypeId:
          type: integer
        leaveTypeName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        reason:
          type: ["string", "null"]
        approverRemark:
          type: ["string", "null"]
        createdAt:
          type: string
    LeaveRequestInput:
      type: object
      required: [leaveTypeId, startDate, endDate]
      properties:
        leaveTypeId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
        isEmergency:
          type: boolean
    LeaveBalance:
      type: object
      required: [leaveTypeId, leaveTypeName, entitled, taken, remaining]
      properties:
        leaveTypeId:
          type: integer
        leaveTypeName:
          type: string
        color:
          type: ["string", "null"]
        entitled:
          type: number
        taken:
          type: number
        remaining:
          type: number

    # ── ATTENDANCE ────────────────────────────────────────────────────────────
...
    ExpenseClaim:
      type: object
      required: [id, employeeId, categoryId, amount, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        approverRemark:
          type: ["string", "null"]
        submittedAt:
          type: string
    ExpenseClaimInput:
      type: object
      required: [categoryId, amount]
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        reason:
          type: string
        fromDate:
          type: string
        toDate:
          type: string
    ExpenseCategory:
      type: object
      required: [id, name, code, active]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        active:
          type: boolean
        maxPerClaim:
          type: ["number", "null"]
        yearlyLimit:
          type: ["number", "null"]
    ExpenseCategoryInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        maxPerClaim:
          type: number
        yearlyLimit:
          type: number

    # ── PERFORMANCE ───────────────────────────────────────────────────────────
...
    ApprovalAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject]
        remark:
          type: string

    # ── TENANT ────────────────────────────────────────────────────────────────
...